refactor(#13): refactored async quadratic solver endpoint
