+++ intelligent-service/src/quadratic_service.py
@@ -1,7 +1,11 @@
 """Quadratic equation solver service"""
 
-import numpy as np
 
+from threading import Thread
+import time
+import numpy as np
+import requests
+import config
 
 def solve_quadratic(coeff_a, coeff_b, coeff_c):
     """Returns the discriminant and the solution of the quadratic equation coefficients"""
@@ -26,3 +30,23 @@ def calculate_discriminant(coeff_a, coeff_b, coeff_c):
     """Returns the discriminant of the given quadratic equation coefficients"""
     discriminant = coeff_b * coeff_b - 4 * coeff_a * coeff_c
     return discriminant
+
+def solve_quadratic_async(task_id, coeff_a, coeff_b, coeff_c):
+    """Solve task in thread"""
+    thread = Thread(target=notify_result, args=(task_id, coeff_a, coeff_b, coeff_c))
+    thread.daemon = True
+    thread.start()
+
+
+
+def notify_result(task_id, coeff_a, coeff_b, coeff_c):
+    """Notify the information system about the solved quadratic equation."""
+    payload = solve_quadratic(coeff_a, coeff_b, coeff_c)
+    delay = int(config.__get_delay())  # pylint: disable=W0212
+    host = config.__get_is_host()  # pylint: disable=W0212
+    port = config.__get_is_port()  # pylint: disable=W0212
+    for i in range(delay):
+        print("Working... {}/{}".format(i + 1, delay))
+        time.sleep(1)
+    requests.post('http://' + host + ":" + port + "/qe/async/notify/" + task_id, json=payload)
+    print("Thread Finished :")
